Inserted cast for compiler for bug6377

+++ libraries/DataFlash/LogFile.cpp
@@ -294,9 +294,9 @@ void DataFlash_Class::Log_Write_RFND(const RangeFinder &rangefinder)
         LOG_PACKET_HEADER_INIT((uint8_t)(LOG_RFND_MSG)),
         time_us       : AP_HAL::micros64(),
         dist1         : rangefinder.distance_cm(0),
-        orient1       : rangefinder.get_orientation(0),
+        orient1       : (uint8_t)rangefinder.get_orientation(0),
         dist2         : rangefinder.distance_cm(1),
-        orient2       : rangefinder.get_orientation(1)
+        orient2       : (uint8_t)rangefinder.get_orientation(1)
     };
     WriteBlock(&pkt, sizeof(pkt));
 }
